+++ ext/Cache/Cell.svelte
@@ -1,7 +1,10 @@
 <script>
   import { Link } from 'svelte-routing';
+  import Checkbox from '../Checkbox.svelte';
+
   export let row;
   export let col;
+  export let onSelect;
 
   const shortenKeys = ['host', 'path'];
   const shorten = (str) => {
@@ -22,8 +25,16 @@
 <div class="Cell">
   {#if col.key === 'edit'}
     <Link to="/cache/entry/{row.id}" data-cy-id="cache-cell-{cyID}"
-      >{`${value.slice(0, 5)}...`}</Link
+      >{`${value.slice(0, 15)}...`}</Link
     >
+  {:else if col.key === 'select'}
+    <Checkbox
+      name="select-row-{row.id}"
+      id="select-row-{row.id}"
+      label=""
+      checked={row.checked}
+      onClick={(e) => onSelect(row.id, e.target.checked)}
+    />
   {:else}
     <span data-cy-id="cache-cell-{cyID}">{value}</span>
   {/if}